Added extra condition on existing profile image

+++ app/Http/Controllers/API/User/UserProfileController.php
@@ -137,7 +137,10 @@ class UserProfileController extends Controller
 
             if ($file_data != null) {
                 Storage::disk('public')->put('user_images/' . $file_name, base64_decode($file_data));
-                Storage::disk('public')->delete('user_images/' . $user->profile_image);
+                if ($user->profile_image != null) {
+                    Storage::disk('public')->delete('user_images/' . $user->profile_image);
+                }
+
             }
             $user->update([
                 'profile_image' => $file_name,
